refactor(auth): 패키지 구조 변경 및 리팩토링
- 멤버, 매개변수 final 작업
- claim name static 멤버로 추출
