+++ app/src/main/java/com/codesoom/assignment/auth/application/AuthenticationService.java
@@ -1,10 +1,10 @@
-package com.codesoom.assignment.session.application;
+package com.codesoom.assignment.auth.application;
 
+import com.codesoom.assignment.auth.application.exception.LoginFailException;
+import com.codesoom.assignment.auth.application.port.in.AuthenticationUseCase;
 import com.codesoom.assignment.common.utils.JwtUtil;
 import com.codesoom.assignment.role.repository.Role;
 import com.codesoom.assignment.role.repository.RoleRepository;
-import com.codesoom.assignment.session.application.port.in.AuthenticationUseCase;
-import com.codesoom.assignment.session.exception.LoginFailException;
 import com.codesoom.assignment.user.repository.User;
 import com.codesoom.assignment.user.repository.UserRepository;
 import io.jsonwebtoken.Claims;
@@ -12,6 +12,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static com.codesoom.assignment.common.utils.JwtUtil.CLAIM_NAME;
 
 @Service
 public class AuthenticationService implements AuthenticationUseCase {
@@ -43,7 +45,7 @@ public class AuthenticationService implements AuthenticationUseCase {
 
     public Long parseToken(final String accessToken) {
         Claims claims = jwtUtil.decode(accessToken);
-        return claims.get("userId", Long.class);
+        return claims.get(CLAIM_NAME, Long.class);
     }
 
     public List<Role> roles(final Long userId) {
