+++ app/src/test/java/com/codesoom/assignment/presentation/AuthenticationProvider.java
@@ -1,8 +1,8 @@
 package com.codesoom.assignment.presentation;
 
+import com.codesoom.assignment.auth.application.exception.InvalidTokenException;
+import com.codesoom.assignment.auth.application.port.in.AuthenticationUseCase;
 import com.codesoom.assignment.role.repository.Role;
-import com.codesoom.assignment.session.application.port.in.AuthenticationUseCase;
-import com.codesoom.assignment.session.exception.InvalidTokenException;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.mock.mockito.MockBean;
