+++ app/src/test/java/com/codesoom/assignment/utils/JwtUtilTest.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.utils;
 
-import com.codesoom.assignment.session.exception.InvalidTokenException;
+import com.codesoom.assignment.auth.application.exception.InvalidTokenException;
 import com.codesoom.assignment.common.utils.JwtUtil;
 import io.jsonwebtoken.Claims;
 import org.junit.jupiter.api.BeforeEach;
